refactor: restructured file for easy maintenance
