+++ backend/services/file_service.py
@@ -1,4 +1,8 @@
 import os
+from flask import Blueprint, jsonify, current_app
+
+def get_case_files_upload_folder():
+    return os.path.join(current_app.root_path, "uploads", "case_files")
 
 def resolve_filename_conflict(directory, original_filename):
     name, ext = os.path.splitext(original_filename)
